# Conflicts:
#	apps/gateway/src/chat/tools/parse-provider-response.ts
#	apps/gateway/src/chat/tools/transform-response-to-openai.ts
#	apps/gateway/src/lib/costs.spec.ts
#	apps/gateway/src/lib/costs.ts

+++ apps/gateway/src/realtime/billing.ts
@@ -96,7 +96,7 @@ export async function getUnsettledRealtimeOrganizationSpend(
 		.select({
 			total: sql<
 				string | null
-			>`sum(coalesce(${log.billingCost}, ${log.cost}::numeric))`,
+			>`sum(coalesce(${log.billingCost}, ${log.cost}::float8::numeric))`,
 		})
 		.from(log)
 		.where(
